created custom hook for getting movie detail
